deployment/requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
gitpython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless
Pillow>=7.1.2
psutil 
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<4.21.5  # https://github.com/ultralytics/yolov5/issues/8012
ultralytics
# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0


# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
streamlit==1.29.0
wget
ffmpeg-python
pyngrok
streamlit_webrtc